branch-data-automation/src/requirements.txt
schedule
python-dotenv
pyodbc
paramiko
requests
smtplib